@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace pipelines_dotnet_core
 {
@@ -31,12 +32,17 @@ namespace pipelines_dotnet_core
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
+           services.Configure<MvcOptions>(options => {
+             options.EnableEndpointRouting = false;
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+           });
+
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        // Note: IHostingEnvironment is Obselete and replaced by IHostEnvironment, so updated to IHostEnvironment
+        public void Configure(IApplicationBuilder app, IHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -53,6 +59,8 @@ namespace pipelines_dotnet_core
             app.UseStaticFiles();
             app.UseCookiePolicy();
 
+            //warning MVC1005: Using 'UseMvc' to configure MVC is not supported while using Endpoint Routing. To continue using 'UseMvc'
+            //, please set 'MvcOptions.EnableEndpointRouting = false' inside 'ConfigureServices'
             app.UseMvc(routes =>
             {
                 routes.MapRoute(
